Push_To_PubSub/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub==2.16.0
